added EOS on GPU
