+++ sph/include/sph/sph.cuh
@@ -1,9 +1,15 @@
 #pragma once
 
 #include "cstone/sfc/box.hpp"
+#ifdef USE_CUDA
+#include "sph/util/cuda_utils.cuh"
+#endif
 
 //! @brief maximum number of neighbors supported in GPU kernels
 #define NGMAX 150
+
+template<class ThrustVec>
+typename ThrustVec::value_type* rawPtr(ThrustVec& p);
 
 namespace sph
 {
@@ -34,6 +40,9 @@ extern void computeAVswitches(size_t, size_t, size_t, Dataset& d, const cstone::
 
 template<class Dataset>
 extern void computeMomentumEnergy(size_t, size_t, size_t, Dataset& d, const cstone::Box<typename Dataset::RealType>&);
+
+template<class Tu, class Trho, class Tp, class Tc>
+extern void computeEOS_HydroStd(size_t, size_t, const Tu*, const Trho*, Tp*, Tc*);
 
 } // namespace cuda
 } // namespace sph
